backend/requirements.txt
fastapi
uvicorn[standard]
pillow
requests
numpy
scikit-learn
sentence-transformers
python-multipart
aiohttp
